Mueve carpetas de mascotas dentro de la carpeta pets
