+++ lib/routes.dart
@@ -9,11 +9,11 @@ import './src/pets/selected_pet_screen.dart';
 import './src/sign_up/user/components/sign_up_details_form.dart';
 import './src/sign_up/user/sign_up_screen.dart';
 import './src/pets/components/add_pets.dart';
-import './src/medical_records/medical_records_screen.dart';
+import './src/pets/medical_records/medical_records_screen.dart';
 import './src/agenda/agenda_screen.dart';
-import './src/medical_records/medical_record_form_screen.dart';
-import './src/vaccionation_records/vaccination_records_form_screen.dart';
-import './src/vaccionation_records/vaccination_records_screen.dart';
+import './src/pets/medical_records/medical_record_form_screen.dart';
+import './src/pets/vaccionation_records/vaccination_records_form_screen.dart';
+import './src/pets/vaccionation_records/vaccination_records_screen.dart';
 import './src/agenda/activity_form_screen.dart';
 import './src/pets/lost_and_found/lost_and_found_map.dart';
 import './src/pets/lost_and_found/found_pet/found_pet_form_step1.dart';
