+++ lib/src/pets/selected_pet_screen.dart
@@ -6,8 +6,8 @@ import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
 import '../../provider/medical_record.dart';
-import '../medical_records/medical_records_screen.dart';
-import '../vaccionation_records/vaccination_records_screen.dart';
+import './medical_records/medical_records_screen.dart';
+import './vaccionation_records/vaccination_records_screen.dart';
 import '../../provider/pets.dart';
 import '../profile/components/profile_menu.dart';
 import './components/pet_pic.dart';
